stagefright: Fix buffer handle retrieval in signalBufferReturned
In commit 3e32878 the stagefright code was restructured to fix
the logic for native handle source, but the change in the
function SurfaceMediaSource::signalBufferReturned was probably
missed.

Try to compare the media buffer handle also to the current native
buffer handle in this function when searching for correspondance.

Change-Id: I352293e525f75dde500ac8e71ee49209710030c3
(cherry picked from commit d7ff18d2c447d911eb25a6bb0ef97528482957ea)